Fix import paths and async error handling after API client refactoring
Fixed build and test failures caused by incorrect import paths and missing async/await:

- Fixed import path in error-handling.ts: '../../shared/error-utils' → '../../../shared/error-utils'
- Fixed import path in react-query-hooks.ts: './api-client' → '../api-client'
- Made handleErrorResponse() async and added await for parseErrorResponse()
- Added await before handleErrorResponse() in api-client.ts

These fixes resolve the module extraction refactoring where api-client was split into
multiple modules in src/lib/api/ but import paths weren't updated correctly.

Build: ✓ Pass
Lint: ✓ 0 errors
Typecheck: ✓ Pass
Tests: ✓ 2416 passing, 5 skipped, 155 todo
